fix(pr-execute): address 5 findings from triage
- ci-status: EmitAddCaseParticipantNode extracted to nodes/accept_invite.py; _resolve_actor_recipients uses actor_participant_index keys (HTTP URLs) not case_participants (bare UUIDs)
- ci-status: accept_invite_tree.py unused imports removed (BTBridge, create_commit_log_entry_tree, _snapshot_with_context)
- test-coverage: add test_to_field_contains_http_actor_urls_not_bare_uuids verifying to= recipients are HTTP actor URLs
- pr-body-format: move Closes #1689 to top, add Changes section
- notes-currency: document invite_actor_to_case disposition=recorded and EmitAddCaseParticipantNode fan-out pattern in case-ledger-authority.md
